requirement.txt
langchain
langchain_community
python-dotenv
openai
fastapi
uvicorn